Add files via upload
A low-down on the history of the inter-State conflict in the Cauvery River basin in India, by mapping all the 100 dams in the region on top of the three river basins. Clicking on the icons gives information of the name of the dam and the year it was built. The Cauvery Basin is shown in green, with Upper Sub Basin depicted in a dark shade, and the Mid Sub Basin and the Lower in progressively lighter hues.
Data: india-wris.nrsc.gov.in | Visualisation: Mahima A Jain | Note: The location of the dam is identified with the nearest town.